# docs/swagger.yaml
definitions:
  controllers.LoginInput:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  controllers.RegisterInput:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
  controllers.bookInput:
    properties:
      judul:
        type: string
      penerbit:
        type: string
      penulis:
        type: string
      tahun_terbit:
        type: integer
    required:
    - judul
    - penerbit
    - penulis
    - tahun_terbit
    type: object
  controllers.commentInput:
    properties:
      deskripsi:
        type: string
      review_id:
        type: integer
      user_id:
        type: integer
    required:
    - deskripsi
    - review_id
    - user_id
    type: object
  controllers.likeInput:
    properties:
      review_id:
        type: integer
      user_id:
        type: integer
    required:
    - review_id
    - user_id
    type: object
...
  models.Books:
    properties:
      created_at:
        type: string
      id:
        type: integer
      judul:
        type: string
      penerbit:
        type: string
      penulis:
        type: string
      tahun_terbit:
        type: integer
      updated_at:
        type: string
    type: object
  models.Comments:
    properties:
      created_at:
        type: string
      deskripsi:
        type: string
      id:
        type: integer
      review_id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  models.Likes:
    properties:
      created_at:
        type: string
      id:
        type: integer
      review_id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
  /book:
    get:
      description: Get a list of book.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Books'
            type: array
      summary: Get all book.
      tags:
      - book
    post:
      description: Creating a new Book.
      parameters:
      - description: the body to create a new Book
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.bookInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Books'
      security:
      - BearerToken: []
      summary: Create New Book.
      tags:
      - book
  /book/{id}:
    delete:
      description: Delete a Book by id.
      parameters:
      - description: Book id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete book.
      tags:
      - book
    get:
      description: Get a book by id.
      parameters:
      - description: Book id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Books'
      summary: Get Book.
      tags:
      - book
    patch:
      description: Update Book by id.
      parameters:
      - description: Book id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update book
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.bookInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Books'
      security:
      - BearerToken: []
      summary: Update Book.
      tags:
      - book
  /comment:
    get:
      description: Get a list of comment.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Comments'
            type: array
      summary: Get all comment.
      tags:
      - comment
    post:
      description: Creating a new Comment.
      parameters:
      - description: the body to create a new Comment
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.commentInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Comments'
      security:
      - BearerToken: []
      summary: Create New Comment.
      tags:
      - comment
  /comment/{id}:
    delete:
      description: Delete a Comment by id.
      parameters:
      - description: Comment id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete comment.
      tags:
      - comment
    patch:
      description: Update Comment by id.
      parameters:
      - description: Comment id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update comment
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.commentInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Comments'
      security:
      - BearerToken: []
      summary: Update Comment.
      tags:
      - comment
  /like:
    get:
      description: Get a list of like.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Likes'
            type: array
      summary: Get all like.
      tags:
      - like
    post:
      description: Creating a new Like.
      parameters:
      - description: the body to create a new Like
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.likeInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Likes'
      security:
      - BearerToken: []
      summary: Create New like.
      tags:
      - like
  /like/{id}:
    delete:
      description: Delete a Like by id.
      parameters:
      - description: Like id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete one like.
      tags:
      - like
  /login:
    post:
      description: Logging in to get jwt token to access admin or user api by roles.
      parameters:
      - description: the body to login a user
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Login as as user.
      tags:
      - Auth
...
  /register:
    post:
      description: registering a user from public access.
      parameters:
      - description: the body to register a user
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.RegisterInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Register a user.
      tags:
      - Auth
...
swagger: "2.0"